# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger NewsApi",
        "description": "News Api",
        "version": "1.0.0"
    },
    "paths": {
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get all new news",
                "operationId": "df53f5fbfbc3a8e5a3fadee520e9b733",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "News Text"
                                            },
                                            "text": {
                                                "type": "string",
                                                "example": "News Text"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "img"
                                            },
                                            "is_active": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "published_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "1900-01-01 12:00:00"
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-21T12:58:45.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-21T12:58:45.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Store a new news",
                "operationId": "e2dbda79b1ec964876c7ae09b6777820",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "News Title"
                                            },
                                            "text": {
                                                "type": "string",
                                                "example": "News Text"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "News Text"
                                                },
                                                "text": {
                                                    "type": "string",
                                                    "example": "News Text"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/news/{post}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news by id",
                "operationId": "95ed817bbdd270fcd3c63046a17b1112",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "news ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "News Text"
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "News Text"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "img"
                                        },
                                        "is_active": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "published_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "1900-01-01 12:00:00"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-21T12:58:45.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-21T12:58:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "News"
                ],
                "summary": "Update news by id",
                "operationId": "dc3b7718a4568977fe1788c9d283207c",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "news ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "title": {
                                                "type": "string",
                                                "example": "Edit News Title"
                                            },
                                            "text": {
                                                "type": "string",
                                                "example": "Edit News Text"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "News Text"
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "News Text"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "img"
                                        },
                                        "is_active": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "published_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "1900-01-01 12:00:00"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-21T12:58:45.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-21T12:58:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Delete news by id",
                "operationId": "ab8cea755cd2f2326e821a50bab49ff2",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "news ID",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "News deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}